Chasing away minor compiler warnings

+++ src/headers/Platform_Windows.hpp
@@ -44,7 +44,7 @@ void Platform::onKeyPress(char pressed)
   if (toSend != nullptr)
   {
     std::string s(toSend->expandsTo);
-    simulateKeyboardInput(strlen(toSend->abbreviation), s);
+    simulateKeyboardInput((int)strlen(toSend->abbreviation), s);
   }
 }
